requirements.txt
gensim
cython
seaborn
openpyxl
xlrd
paramiko
pyICU
absl-py>=0.6.1
abydos>=0.3.6
appnope>=0.1.0
astor>=0.7.1
astroid>=2.1.0
autopep8>=1.4.3
backcall>=0.1.0
bleach>=3.0.2
cachetools>=3.0.0
certifi>=2018.11.29
chardet>=3.0.4
cycler>=0.10.0
cyrtranslit>=0.4
decorator>=4.3.0
defusedxml>=0.5.0
entrypoints>=0.2.3
# Feast>=0.1.0
Fuzzy>=1.2.2
fuzzywuzzy>=0.17.0
gast>=0.2.0
google-api-core>=1.7.0
google-api-python-client>=1.7.7
google-auth>=1.6.2
google-auth-httplib2>=0.0.3
google-cloud-bigquery>=1.8.1
google-cloud-core>=0.29.1
google-cloud-storage>=1.13.2
google-resumable-media>=0.3.2
googleapis-common-protos>=1.5.5
grpcio>=1.17.1
h5py>=2.9.0
httplib2>=0.12.0
idna>=2.8
ipykernel>=5.1.0
ipython>=7.2.0
ipython-genutils>=0.2.0
ipywidgets>=7.4.2
isort>=4.3.4
jedi>=0.13.2
jellyfish>=0.7.1
Jinja2>=2.10
jsonschema>=2.6.0
jupyter>=1.0.0
jupyter-client>=5.2.4
jupyter-console>=6.0.0
jupyter-contrib-core>=0.3.3
jupyter-core>=4.4.0
jupyter-nbextensions-configurator>=0.4.1
kaggle>=1.5.1.1
Keras-Applications>=1.0.6
Keras-Preprocessing>=1.0.5
kiwisolver>=1.0.1
lazy-object-proxy>=1.3.1
Markdown>=3.0.1
MarkupSafe>=1.1.0
matplotlib>=3.0.2
mccabe>=0.6.1
mistune>=0.8.4
mlperf-compliance>=0.0.10
mock>=2.0.0
nbconvert>=5.4.0
nbformat>=4.4.0
networkx>=1.8.1
nltk>=3.4
notebook>=5.7.4
numpy>=1.15.4
oauth2client>=4.1.3
pandas>=0.23.4
pandocfilters>=1.4.2
parso>=0.3.1
pbr>=5.1.1
pexpect>=4.6.0
pickleshare>=0.7.5
prometheus-client>=0.5.0
prompt-toolkit>=2.0.7
protobuf>=3.6.1
psutil>=5.4.8
ptyprocess>=0.6.0
py-cpuinfo>=4.0.0
py-stringmatching>=0.4.0
pyasn1>=0.4.5
pyasn1-modules>=0.2.3
pycodestyle>=2.4.0
Pygments>=2.3.1
pylint>=2.2.2
pymongo>=3.7.2
pyparsing>=2.3.1
python-dateutil>=2.7.5
python-Levenshtein>=0.12.0
python-slugify>=2.0.1
pytz>=2018.7
pyxDamerauLevenshtein>=1.5.2
PyYAML>=3.13
pyzmq>=17.1.2
qtconsole>=4.4.3
requests>=2.21.0
rsa>=4.0
scikit-learn>=0.20.2
scipy>=1.2.0
Send2Trash>=1.5.0
singledispatch>=3.4.0.3
six>=1.12.0
sklearn>=0.0
sonnet>=0.1.6
strsim>=0.0.3
tensorboard>=1.12.2
tensorflow>=1.12.0
# tensorflow-ranking>=0.1.0
termcolor>=1.1.0
terminado>=0.8.1
testpath>=0.4.2
textblob>=0.15.2
textdistance>=3.0.3
tornado>=5.1.1
tqdm>=4.28.1
traitlets>=4.3.2
transliterate>=1.10.2
typed-ast>=1.1.1
typing>=3.6.6
Unidecode>=1.0.23
uritemplate>=3.0.0
urllib3>=1.22
wcwidth>=0.1.7
webencodings>=0.5.1
Werkzeug>=0.14.1
widgetsnbextension>=3.4.2
wrapt>=1.10.11
xgboost>=0.81
